Modularize projection rewriting code from ProjectionPullUpJoin rule (and fix compiler warnings
