+++ optd-datafusion-repr/src/rules/filter_pushdown.rs
@@ -9,7 +9,6 @@
 //! further down the query plan whenever it is possible to do so.
 
 use std::collections::HashMap;
-use std::result;
 
 use optd_core::rules::{Rule, RuleMatcher};
 use optd_core::{optimizer::Optimizer, rel_node::RelNode};
@@ -24,11 +23,15 @@ define_rule!(
     (Filter, child, [cond])
 );
 
-/// Projection and filter are commutable when TODO
+/// Datafusion only pushes filter past project when the project does not contain
+/// volatile (i.e. non-deterministic) expressions that are present in the filter
+/// Calcite only checks if the projection contains a windowing calculation
+/// We are checking neither of those things here right now
 fn filter_project_transpose(
-    child: RelNode<OptRelNodeTyp>,
-    cond: RelNode<OptRelNodeTyp>,
+    _child: RelNode<OptRelNodeTyp>,
+    _cond: RelNode<OptRelNodeTyp>,
 ) -> Vec<RelNode<OptRelNodeTyp>> {
+    // let old_proj = LogicalProjection::from_rel_node(child.into()).unwrap();
     vec![]
 }
 
@@ -50,7 +53,7 @@ fn apply_filter_pushdown(
     FilterPushdownRulePicks { child, cond }: FilterPushdownRulePicks,
 ) -> Vec<RelNode<OptRelNodeTyp>> {
     // Push filter down one node
-    let result_from_this_step = match child.typ {
+    let mut result_from_this_step = match child.typ {
         OptRelNodeTyp::Projection => filter_project_transpose(child, cond),
         OptRelNodeTyp::Filter => todo!(), // @todo filter merge rule? Should we do that here?
         OptRelNodeTyp::Scan => todo!(),   // @todo Why doesn't our sort node have a predicate field?
@@ -58,41 +61,33 @@ fn apply_filter_pushdown(
         OptRelNodeTyp::Sort => filter_sort_transpose(child, cond),
         OptRelNodeTyp::Agg => todo!(),
         OptRelNodeTyp::Apply(_) => todo!(),
-        OptRelNodeTyp::EmptyRelation => todo!(),
-        OptRelNodeTyp::Limit => todo!(),
         _ => vec![],
     };
 
     // Apply rule recursively
-    if let Some(new_node) = result_from_this_step.first() {
+    if let Some(new_node) = result_from_this_step.first_mut() {
         // For all the children in our result,
-        new_node.children = new_node
-            .children
-            .into_iter()
-            .map(|child| {
-                if child.typ == OptRelNodeTyp::Filter {
-                    // If this node is a filter, apply the rule again to this node!
-                    let child_as_filter = LogicalFilter::from_rel_node(child.clone()).unwrap();
-                    let childs_child = child_as_filter.child().into_rel_node().as_ref().clone();
-                    let childs_cond = child_as_filter.cond().into_rel_node().as_ref().clone();
-                    // @todo: make this iterative?
-                    let result = apply_filter_pushdown(
-                        _optimizer,
-                        FilterPushdownRulePicks {
-                            child: childs_child,
-                            cond: childs_cond,
-                        },
-                    );
-                    // If we got a result, that is the replacement for this child
-                    if let Some(&new_child) = result.first() {
-                        return new_child.into();
-                    }
+        for child in new_node.children.iter_mut() {
+            if child.typ == OptRelNodeTyp::Filter {
+                // If this node is a filter, apply the rule again to this node!
+                let child_as_filter = LogicalFilter::from_rel_node(child.clone()).unwrap();
+                let childs_child = child_as_filter.child().into_rel_node().as_ref().clone();
+                let childs_cond = child_as_filter.cond().into_rel_node().as_ref().clone();
+                // @todo: make this iterative?
+                let result = apply_filter_pushdown(
+                    _optimizer,
+                    FilterPushdownRulePicks {
+                        child: childs_child,
+                        cond: childs_cond,
+                    },
+                );
+                // If we got a result, that is the replacement for this child
+                if let Some(&new_child) = result.first().as_ref() {
+                    *child = new_child.to_owned().into();
                 }
-                // Otherwise, if there was no result from rule application or this is not a filter,
-                // return the old child unmodified.
-                child
-            })
-            .collect();
+            }
+            // Otherwise, if there was no result from rule application or this is not a filter, do not modify the child
+        }
     }
 
     result_from_this_step
