+++ optd-datafusion-repr/src/rules/joins.rs
@@ -256,35 +256,6 @@ define_rule!(
     )
 );
 
-struct ProjectionMapping {
-    forward: Vec<usize>,
-    _backward: Vec<Option<usize>>,
-}
-
-impl ProjectionMapping {
-    pub fn build(mapping: Vec<usize>) -> Option<Self> {
-        let mut backward = vec![];
-        for (i, &x) in mapping.iter().enumerate() {
-            if x >= backward.len() {
-                backward.resize(x + 1, None);
-            }
-            backward[x] = Some(i);
-        }
-        Some(Self {
-            forward: mapping,
-            _backward: backward,
-        })
-    }
-
-    pub fn projection_col_refers_to(&self, col: usize) -> usize {
-        self.forward[col]
-    }
-
-    pub fn _original_col_maps_to(&self, col: usize) -> Option<usize> {
-        self._backward[col]
-    }
-}
-
 fn apply_projection_pull_up_join(
     optimizer: &impl Optimizer<OptRelNodeTyp>,
     ProjectionPullUpJoinPicks {
@@ -294,72 +265,21 @@ fn apply_projection_pull_up_join(
         cond,
     }: ProjectionPullUpJoinPicks,
 ) -> Vec<RelNode<OptRelNodeTyp>> {
-    let list = ExprList::from_rel_node(Arc::new(list)).unwrap();
-
-    fn compute_column_mapping(list: ExprList) -> Option<ProjectionMapping> {
-        let mut mapping = vec![];
-        for expr in list.to_vec() {
-            let col_expr = ColumnRefExpr::from_rel_node(expr.into_rel_node())?;
-            mapping.push(col_expr.index());
-        }
-        ProjectionMapping::build(mapping)
-    }
-
-    let Some(mapping) = compute_column_mapping(list.clone()) else {
-        return vec![];
-    };
-
-    fn rewrite_condition(
-        cond: Expr,
-        mapping: &ProjectionMapping,
-        left_schema_size: usize,
-        projection_schema_size: usize,
-    ) -> Expr {
-        if cond.typ() == OptRelNodeTyp::ColumnRef {
-            let col = ColumnRefExpr::from_rel_node(cond.into_rel_node()).unwrap();
-            let idx = col.index();
-            if idx < projection_schema_size {
-                let col = mapping.projection_col_refers_to(col.index());
-                return ColumnRefExpr::new(col).into_expr();
-            } else {
-                let col = col.index();
-                return ColumnRefExpr::new(col - projection_schema_size + left_schema_size)
-                    .into_expr();
-            }
-        }
-        let expr = cond.into_rel_node();
-        let mut children = Vec::with_capacity(expr.children.len());
-        for child in &expr.children {
-            children.push(
-                rewrite_condition(
-                    Expr::from_rel_node(child.clone()).unwrap(),
-                    mapping,
-                    left_schema_size,
-                    projection_schema_size,
-                )
-                .into_rel_node(),
-            );
-        }
-
-        Expr::from_rel_node(
-            RelNode {
-                typ: expr.typ.clone(),
-                children,
-                data: expr.data.clone(),
-            }
-            .into(),
-        )
-        .unwrap()
-    }
-
     let left = Arc::new(left.clone());
     let right = Arc::new(right.clone());
 
+    let list = ExprList::from_rel_node(Arc::new(list)).unwrap();
+
+    let projection = LogicalProjection::new(PlanNode::from_group(left.clone()), list.clone());
+
+    let Some(mapping) = projection.compute_column_mapping() else {
+        return vec![];
+    };
+
     // TODO(chi): support capture projection node.
-    let projection =
-        LogicalProjection::new(PlanNode::from_group(left.clone()), list.clone()).into_rel_node();
     let left_schema = optimizer.get_property::<SchemaPropertyBuilder>(left.clone(), 0);
-    let projection_schema = optimizer.get_property::<SchemaPropertyBuilder>(projection.clone(), 0);
+    let projection_schema =
+        optimizer.get_property::<SchemaPropertyBuilder>(projection.into_rel_node().clone(), 0);
     let right_schema = optimizer.get_property::<SchemaPropertyBuilder>(right.clone(), 0);
     let mut new_projection_exprs = list.to_vec();
     for i in 0..right_schema.len() {
@@ -370,9 +290,8 @@ fn apply_projection_pull_up_join(
         LogicalJoin::new(
             PlanNode::from_group(left),
             PlanNode::from_group(right),
-            rewrite_condition(
+            mapping.rewrite_condition(
                 Expr::from_rel_node(Arc::new(cond)).unwrap(),
-                &mapping,
                 left_schema.len(),
                 projection_schema.len(),
             ),
